marketplace/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Deprecated==1.2.13
Django==2.2.19
django-cors-headers==3.2.0
django-js-asset==2.0.0
django-mptt==0.13.4
djangorestframework==3.12.4
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
kombu==5.2.4
nose==1.3.7
packaging==21.3
prompt-toolkit==3.0.29
psycopg2-binary==2.8.6
pyparsing==3.0.9
pytz==2022.1
redis==4.3.3
requests==2.28.0
six==1.16.0
SQLAlchemy==1.4.37
sqlparse==0.4.2
typing_extensions==4.2.0
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.8.0
